Refactoring glamorous-ified NavItem component
